// src/redux/profiles/enums.ts
export enum Gender {
  Male = "M",
  Female = "F",
}

export enum Status {
  Married = "married",
  Single = "single",
  Divorced = "divorced",
  Widowed = "widowed",
}

export enum LeadershipLevel {
  SeniorPastor = "seniorpastor",
  NetworkHead = "networkhead",
  TwoEightEight = "288",
  ThreeFourFiveSix = "3456",
  TwentyK = "20K",
  Multitudes = "multitudes",
  NetworkChurch = "networkchurch",
  Visitors = "visitors",
}